MINOR: parsers: add backend server-state-file-name attribute
Upstream docs: https://cbonte.github.io/haproxy-dconv/2.2/configuration.html#4.2-server-state-file-name

It felt simple enough to just leave it as a word, instead of a custom parser that contains a separate bool for "use-backend-name", which is mutually exclusive with file path.

This is currently cut off of v3.1.0 for our internal needs.